Moved socketcan.cpp and .hpp to Common folder
- socketcan.cpp in Common/src
- socketcan.hpp in Common/include
- Added Display module can frame in candb.hpp
- Updated SWdesign diagram
