+++ Common/include/candb.hpp
@@ -8,3 +8,12 @@ struct UserInputCanFrame {
   unsigned blinkers: 2;
 };
 
+struct DisplayCanFrame
+{
+  unsigned frame_counter: 4;
+  unsigned ignition: 2;
+  unsigned gear_select: 4;
+  unsigned speed: 8;
+  unsigned rpm: 8;
+  unsigned blinkers: 2;
+};
