--- UserInput/include/canSender.hpp
@@ -1,14 +1,0 @@
-#include <fstream>
-
- struct SentValues
-    {
-        int8_t throttle{};
-        int8_t brake{};
-        int8_t gear{};
-    };
-
-class ValuesToSent
-{
-public:
-   
-};
